+++ src/api/index.ts
@@ -14,4 +14,5 @@ export * from './consume-chain.api';
 export * from './returning-flow-rate.api';
 export * from './system.api';
 export * from './verify.api';
+export * from './actuator.api';
 export * from './metadata.api';